cpio: restructure package, add documentation.
