+++ pkg/uio/reader.go
@@ -5,6 +5,7 @@
 package uio
 
 import (
+	"bytes"
 	"io"
 	"io/ioutil"
 	"math"
@@ -30,4 +31,24 @@ func ReadAll(r io.ReaderAt) ([]byte, error) {
 // Reader generates a Reader from a ReaderAt.
 func Reader(r io.ReaderAt) io.Reader {
 	return io.NewSectionReader(r, 0, math.MaxInt64)
+}
+
+// ReaderAtEqual compares the contents of r1 and r2.
+func ReaderAtEqual(r1, r2 io.ReaderAt) bool {
+	var c, d []byte
+	var err error
+	if r1 != nil {
+		c, err = ReadAll(r1)
+		if err != nil {
+			return false
+		}
+	}
+
+	if r2 != nil {
+		d, err = ReadAll(r2)
+		if err != nil {
+			return false
+		}
+	}
+	return bytes.Equal(c, d)
 }
